+++ src/LayoutProfilePopups.hpp
@@ -8,13 +8,21 @@
 namespace pause_menu_studio {
 class LayoutNamePopup final : public geode::Popup {
 public:
-    static LayoutNamePopup* create(geode::Function<void(std::string)> callback);
+    static LayoutNamePopup* create(
+        geode::Function<void(std::string)> callback,
+        std::string title = "Save layout",
+        std::string initialValue = ""
+    );
 
 private:
     geode::TextInput* m_input = nullptr;
     geode::Function<void(std::string)> m_callback;
 
-    bool init(geode::Function<void(std::string)> callback);
+    bool init(
+        geode::Function<void(std::string)> callback,
+        std::string title,
+        std::string initialValue
+    );
     void onSave(cocos2d::CCObject*);
 };
 
@@ -24,6 +32,8 @@ public:
         std::vector<std::string> names,
         std::string activeName,
         geode::Function<void(std::string)> applyCallback,
+        geode::Function<bool(std::string, std::string)> renameCallback,
+        geode::Function<bool(std::string, std::string)> duplicateCallback,
         geode::Function<void(std::string)> deleteCallback
     );
 
@@ -31,6 +41,8 @@ private:
     std::vector<std::string> m_names;
     std::string m_activeName;
     geode::Function<void(std::string)> m_applyCallback;
+    geode::Function<bool(std::string, std::string)> m_renameCallback;
+    geode::Function<bool(std::string, std::string)> m_duplicateCallback;
     geode::Function<void(std::string)> m_deleteCallback;
     geode::ScrollLayer* m_scroll = nullptr;
 
@@ -38,10 +50,14 @@ private:
         std::vector<std::string> names,
         std::string activeName,
         geode::Function<void(std::string)> applyCallback,
+        geode::Function<bool(std::string, std::string)> renameCallback,
+        geode::Function<bool(std::string, std::string)> duplicateCallback,
         geode::Function<void(std::string)> deleteCallback
     );
     void rebuildList();
     void onApply(cocos2d::CCObject* sender);
+    void onRename(cocos2d::CCObject* sender);
+    void onDuplicate(cocos2d::CCObject* sender);
     void onDelete(cocos2d::CCObject* sender);
 };
 }